refactor: Modularize methods (#19)
